Extra copying of opponent ufuns
Some agents set the reserved value of the opponent_ufun while acting. In
the competition, we are careful to always give these agents copies of
the opponent ufun and to not depend on this setting. Nevertheless,
another user in the future may just pass the opponent_ufun directly
without copying to the privat_info which will allow these agents to
actually change the reserved value of the opponent directly. This will
obviously be a bug which is hard to discover. To avoid this possibility,
we explicitly deep-copy the opponent ufun on_preferences_change in these
agents.